landing page and sign in page changes
