+++ extension/app/(auth)/auth/extension-callback/page.tsx
