+++ extension/app/(auth)/auth/layout.tsx
@@ -1,0 +1,3 @@
+export default function AuthLayout({ children }: { children: React.ReactNode }) {
+  return <>{children}</>
+}
