+++ extension/app/(main)/layout.tsx
@@ -1,0 +1,14 @@
+import { Navbar } from '@/app/components/Navbar'; // Adjust path if needed
+
+export default function MainLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <>
+      <Navbar />
+      <main>{children}</main>
+    </>
+  );
+}
